Refactor : Redux사용해서 코드 변경 완료!
