+++ package.json
@@ -9,9 +9,11 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.1",
+    "react-router-dom": "^6.13.0",
     "react-scripts": "5.0.1",
     "react-uuid": "^2.0.0",
     "redux": "^4.2.1",
+    "styled-components": "^6.0.0-rc.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
