+++ src/components/IsDoneButton.jsx
@@ -1,6 +1,12 @@
-const IsDoneButton = ({ item, removeFunction, clickToggleButtonHandler }) => {
+const IsDoneButton = ({
+  item,
+  removeFunction,
+  clickToggleButtonHandler,
+  detailFunction,
+}) => {
   return (
     <div key={item.id} className="todo-container">
+      <button onClick={() => detailFunction(item.id)}>상세보기</button>
       <h2 className="todo-title">{item.title}</h2>
       <div>{item.content}</div>
       <div className="button-set">
