+++ src/shared/Router.js
@@ -1,0 +1,16 @@
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Detail from "../pages/Detail";
+import Home from "../pages/Home";
+
+const Router = () => {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="detail/:id" element={<Detail />} />
+      </Routes>
+    </BrowserRouter>
+  );
+};
+
+export default Router;
